feat(client, server): add support for Novu notification center
Add support for Novu notification center to the client and server. In the client, the `@novu/notification-center` package was added to the dependencies, and the `Nav` component was updated to include a `NotificationBell` component from the package. In the server, the `@novu/node` package was added to the dependencies, and the `novu` instance was created with the API key. The `create` and `addSubscribers` methods were called on the `topics` object to create a new topic and add a subscriber to it. The `trigger` method was called on the `novu` instance to trigger a notification when a new reply is added.